misc tidyings from another branch
