+++ test/js/bun/io/bun-write.test.js
@@ -273,7 +273,7 @@ const IS_UV_FS_COPYFILE_DISABLED =
 
   it("Bun.file -> Response", async () => {
     using tmpbase = tempDir("bun-file-to-response", {});
-    using server = exampleSite("https");
+    await using server = exampleSite("https");
     // ensure the file doesn't already exist
     try {
       fs.unlinkSync(tmpbase + "fetch.js.out");
