+++ test/js/bun/test/parallel/test-http-7480-should-not-emit-continue-event.ts
@@ -1,7 +1,7 @@
 import { createTest, exampleSite } from "node-harness";
 import https from "node:https";
 const { expect } = createTest(import.meta.path);
-using server = exampleSite();
+const server = exampleSite();
 let receivedContinue = false;
 const req = https.request(server.url, { ca: server.ca, headers: { "accept-encoding": "identity" } }, res => {
   let data = "";
