+++ test/js/bun/test/parallel/test-http-can-send-brotli-from-Server-and-receive-with-Client.ts
@@ -7,8 +7,8 @@ const { expect } = createTest(import.meta.path);
 
 await using server = createServer((req, res) => {
   expect(req.url).toBe("/hello");
-  res.writeHead(200);
   res.setHeader("content-encoding", "br");
+  res.writeHead(200);
 
   const inputStream = new stream.Readable();
   inputStream.push("Hello World");
