+++ test/js/bun/test/parallel/test-http-do-https-request-fails.ts
@@ -1,7 +1,7 @@
 import { createTest, exampleSite } from "node-harness";
 import http from "node:http";
 const { expect } = createTest(import.meta.path);
-using server = exampleSite("https");
+await using server = exampleSite("https");
 expect(() => http.request(server.url)).toThrow(TypeError);
 expect(() => http.request(server.url)).toThrow({
   code: "ERR_INVALID_PROTOCOL",
