+++ test/js/bun/test/parallel/test-http-get-can-use-Agent.ts
@@ -6,5 +6,5 @@ const agent = new http.Agent();
 const { promise, resolve } = Promise.withResolvers();
 http.get({ agent, hostname: "google.com" }, resolve);
 const response = await promise;
-expect(response.req.port).toBe(80);
+expect(response.req.agent.defaultPort).toBe(80);
 expect(response.req.protocol).toBe("http:");
