+++ test/js/bun/test/parallel/test-http-request-has-the-correct-options.ts
@@ -5,5 +5,5 @@ const { expect } = createTest(import.meta.path);
 const { promise, resolve } = Promise.withResolvers();
 http.request("http://google.com/", resolve).end();
 const response = await promise;
-expect(response.req.port).toBe(80);
+expect(response.req.agent.defaultPort).toBe(80);
 expect(response.req.protocol).toBe("http:");
