+++ test/js/bun/test/parallel/test-https-get-can-use-Agent.ts
@@ -6,5 +6,5 @@ const agent = new https.Agent();
 const { promise, resolve } = Promise.withResolvers();
 https.get({ agent, hostname: "google.com" }, resolve);
 const response = await promise;
-expect(response.req.port).toBe(443);
+expect(response.req.agent.defaultPort).toBe(443);
 expect(response.req.protocol).toBe("https:");
